calculating mass

+++ src/main/resources/db/migration/V6__create_material_stock_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE material_stock (
     signature varchar(100) NOT NULL,
     profile varchar(100) NOT NULL,
     material varchar(100) NOT NULL,
-    profile_length double NOT NULL,
+    profile_length DECIMAL(10, 2) NOT NULL,
     qty int NOT NULL,
     available_qty int NOT NULL,
     material_type varchar(100) NOT NULL,
